chore - fix data destructuring & move query hook
