+++ src/api/board/index.ts
@@ -3,9 +3,7 @@ import { http } from '../core';
 export const getMeetingListRegisteredByMe = async (
   params?: req.BoardListRegisteredByMe
 ): Promise<res.BoardListRegisteredByMe> => {
-  const {
-    data: { data },
-  } = await http.get('api/board/me', {
+  const { data } = await http.get('api/board/me', {
     params,
   });
   return data;
