--- src/hooks/api/borad/index.ts
@@ -1,17 +1,0 @@
-import { getMeetingListRegisteredByMe } from 'src/api/board';
-import { queryKey } from 'src/constants/queryKey';
-
-import { useCoreInfiniteQuery } from '../core';
-
-export const useGetMeetingListRegisteredByMeQuery = (
-  params?: req.BoardListRegisteredByMe
-) => {
-  return useCoreInfiniteQuery(
-    queryKey.meetingListRegisteredByMe,
-    ({ pageParam = 0 }) =>
-      getMeetingListRegisteredByMe({ ...params, page: pageParam }),
-    {
-      getNextPageParam: ({ pageable: { pageNumber } }) => pageNumber + 1,
-    }
-  );
-};
